Easy_cartoonify/requirements.txt
torch
torchvision
pillow
numpy
opencv-python
